Rename adapted to stream_adapter
