+++ p2p/src/net/mock/transport/stream_adapter/noise.rs
