+++ p2p/src/net/mock/transport/stream_adapter/traits.rs
